### 2022-04-01 #4292 @editorialbot
[PRE REVIEW]: Handprint: a program to explore and compare major cloud-based services for handwritten text recognition
**Submitting author:** @mhucka (<a href="http://orcid.org/0000-0001-9105-5960">Michael Hucka</a>)
**Repository:** https://github.com/caltechlibrary/handprint
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.5.6
