### 2022-04-01 #4292 @kyleniemeyer
@openjournals/dev any idea why editorialbot isn't able to see the Python files in the repo? I can see plenty of code here: https://github.com/caltechlibrary/handprint/tree/main/handprint
