### 2022-04-01 #4292 @xuanxu
> any idea why editorialbot isn't able to see the Python files in the repo? I can see plenty of code here: https://github.com/caltechlibrary/handprint/tree/main/handprint

The analyzed branch was `joss-paper`, which doesn't include Python files. To run the code analysis on `main` the simplest way is set branch to main, then run `check repository`, and then set the branch back to `joss-paper`.
