### 2022-04-01 #4292 @mhucka
By way of explanation, I followed the "submission requirements" section at https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements where it states:

> Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default).

My interpretation of this was: code & paper must be in the same repo, but it's okay to put the paper in a separate branch. Now, it seems to me the instructions are ambiguous about whether the code should or shouldn't be in that same branch. My reasoning for my current arrangement (i.e., paper in `joss-paper` branch _without_ code) is that there wouldn't be much point in creating a branch _for the paper_ if that branch also contained code.

My apologies if that was a mistake!
