### 2022-04-17 #4292 @mhucka
Despite the 2500+ people in that list, no one seems to have handwritten text recognition specifically listed. The best I could come up with is some guessing based on who has both Python as preferred language and some kind of text analysis/processing.  Here they are in no particular order:
* rlskoeser
* mbod
* cmaimone
* baileythegreen
* adammichaelwood
* step21

If none of them are available, I can do another round by expanding my search or thinking about people outside of that list.
