### 2022-04-17 #4292 @baileythegreen
@danielskatz, I'd be happy to take a look and see if this is something I could review (I'm unsure how related it will prove to be to anything I've done); however, I am on holidays until 27 May, so can not look at it before then.
