### 2022-04-17 #4292 @step21
@danielskatz I would be happy to review it. I haven't worked on image recognition myself yet, but on NLP and am getting more and more into image recognition (one subset of which is handwriting recognition).
