### 2022-04-20 #4292 @mhucka
@rlskoeser @step21 Thank you for being willing to review this!

With respect to accounts: since the software's entire purpose is to make it possible to use cloud-based HTR services, I haven't come up with a way to avoid needing accounts. I wrote up hopefully clear instructions about how to get free accounts on all 3 services, at https://github.com/caltechlibrary/handprint/wiki – free accounts are reasonably generous at all the services and should certainly be enough to try things out.

For the review, maybe getting just one set up would be enough, to save you time?

If worse comes to worse, I could create accounts for reviewers and send you the credentials somehow.
