### 2022-05-04 #4328 @rlskoeser
@danielskatz question about the license: it's described in the readme as a "a BSD/MIT type license" but I don't think it's actually any of the licenses on the OSI list.

Here's the full description in the readme:
> Software produced by the Caltech Library is Copyright © 2021–2022 California Institute of Technology. This software is freely distributed under a BSD/MIT type license.

Please advise.
