### 2022-05-04 #4328 @danielskatz
It's almost https://opensource.org/licenses/BSD-3-Clause but not quite, and not quite isn't good enough, as it's not an OSI-approved license then.

@mhucka, can you remove line 2 ("All rights not granted herein are expressly reserved by Caltech.") from the license files?

And change the readme to say that this uses the 3-Clause BSD License, rather than "a BSD/MIT type license"?
