### 2022-05-05 #4328 @mhucka
Well, hmmm, this _was_ handed down to us by the institute's IP devision, but nevertheless, I'm asking if this policy can be revisited. I'll apprise people here of the answer once I know something.
