### 2022-06-22 #4328 @mhucka
@danielskatz At precisely 5:01 pm today, I finally received an email in answer to the license question. The answer is, yes, for this project (and _only_ this project :-)) we can change the license to be straight BSD. I will update the license file and any associated files and issue a new release of the software.  It will not take long – hopefully tomorrow or Thursday at the latest.
