### 2022-06-22 #4328 @danielskatz
👋 @step21 & @rlskoeser, sorry for the delay - we might have caught the license issue earlier.

But in any case, we should now proceed with (continue) the review of the software.  I look forward to seeing any more issue you find, or seeing the lack of them. 🙂
