+++ setup.cfg
@@ -8,7 +8,7 @@
 
 [metadata]
 name          = handprint
-version       = 1.5.6
+version       = 1.6.0
 description   = Run handwritten text recognition services on images of documents
 author        = Michael Hucka
 author_email  = mhucka@caltech.edu
