+++ CITATION.cff
@@ -12,7 +12,7 @@ repository-code: "https://github.com/caltechlibrary/handprint"
 title: "Handprint: Run handwritten text recognition services on images of documents"
 date-released: 2022-06-24
 version: "1.6.0"
-doi: 10.22002/20059
+doi: 10.22002/20207
 keywords: 
   - handwritten text recognition
   - optical character recognition
