### 2022-06-25 6d98c0c
Remove somewhat out-of-date note about recent changes
