### 2022-06-25 #4328 @mhucka
I've released version 1.6.0 of the software, with a proper unmodified BSD 3-clause license.
