### 2022-07-22 #4328 @rlskoeser
@mhucka I'm glad you got the permission to change the license. I see that you edited both `LICENSE` and `LICENSE.html` but GitHub doesn't seem to recognize the html version:

![Screen Shot 2022-07-22 at 5 09 33 PM](https://user-images.githubusercontent.com/691231/180569106-2e8605c6-33b2-4ea4-b70a-ca00f0cea488.png)
![Screen Shot 2022-07-22 at 5 09 38 PM](https://user-images.githubusercontent.com/691231/180569115-045ba1d7-1036-4b8d-a5c4-031c6598dc98.png)

I wasn't sure if this merited opening an issue on the repo, but thought I would point it out since it seems like it could cause confusion.
