### 2022-07-24 #4328 @mhucka
@rlskoeser Thanks for spotting that. I guess the html version is not really needed anyway, so I'll delete it right now. Thanks again!
