### 2022-07-24 1b84db7
Remove unneeded html version of license file
As [pointed out by
@rlskoeser](https://github.com/openjournals/joss-reviews/issues/4328#issuecomment-1192938621),
the LICENSE.html file confuses GitHub's license navigator. I don't
think I use it for anything anymore, so let's just delete it.
