### 2022-07-26 #4328 @step21
I had some trouble getting testing access at least with google. Azure worked better but did not test yet. Haven't tried Amazon yet. If you would have any test credentials, this could help as I think I might have to sign up for a whole new Google account to get testing credit again.
