### 2022-07-26 caltechlibrary/handprint#38 @step21
JOSS review - paper / state of the field
As far as I can tell there is no review of the "state of the field" in the paper. Or do no other, somewhat similar tools exist? Alternatively, you could also compare to existing solutions / use of the services without tools.
And, not totally sure, but a lot of the usage description might be better relegated to the Readme or documentation, which the paper is not supposed to be I think.
