### 2022-07-26 caltechlibrary/handprint#39 @step21
JOSS Paper - Documentation - Example usage
Hi, I think ideally you could add exmaple usage instructions also to the wiki. This would be easier for users, so after the installation part they could directly test something, without looking for the documentation site. (maybe even directly with an image from the tests or other freely available image so that users can test without having their data at the ready)
