### 2022-07-26 #4328 @step21
Issue re paper that there is not comparison to the field and maybe too much much usage description in the paper. https://github.com/caltechlibrary/handprint/issues/38
