### 2022-07-26 #4328 @step21
Also I would prefer if a direct usage example would also be present in the wiki, not only in the documentation site. https://github.com/caltechlibrary/handprint/issues/39
