### 2022-07-29 caltechlibrary/handprint#38 @rlskoeser
I agree with @step21 — no state of the field review, and too much detail on the documentation.

I think the details of functionality, configuration and usage are beyond the kind of high-level overview that is really needed here. For example, the list of specific image formats and ways to specify images seemed irrelevant to me. Stating that the tool supports a wide range of common image formats and that there are multiple ways to provide images is probably enough here.

I am curious about the state of the field — are there any other tools like this, whether for HTR or OCR? I know about a student project on OCR benchmarking a few years ago because the student won the senior thesis prize from the Center for Digital Humanities at Princeton, where I work. Here's the write up, in case it's at all relevant: https://cdh.princeton.edu/updates/meet-the-2021-cdh-senior-thesis-prize-honorees-william-ughetta/

Are there other similar comparison tools, or do people typically stick with one API that they already have access to whether or not it is best suited to a particular set of content? Do these services not have that much traction, or are people using other tools to access them without running comparisons?
