### 2022-07-29 caltechlibrary/handprint#38 @rlskoeser
I am a little curious about the statement of need — is this tool primarily for comparing the APIs, or do you expect people to use handprint to compare the APIs and make a choice, and then continue to use handprint for their actual workflows?

I'm asking because I was struck by the last sentence in the statement of need section, that "if desired" people could use handprint for automated workflows. Do you know if people are doing so? Would handprint be as good as an API-specific tool, or would there be some kind of tradeoff in using handprint vs something else?

I'm not sure how much you need to clarify this for the paper, but it would be stronger if you did.
