### 2022-07-29 #4328 @rlskoeser
I agree with @step21 about the paper. Added some comments on https://github.com/caltechlibrary/handprint/issues/38 to elaborate on what could be condensed/simplified and what I'm interested to learn more about that isn't currently included.
