### 2022-07-29 #4328 @step21
@rlskoeser did you do any functionality testing yet? My main problem with that is that I did not yet manage to get a google account with free credit - only for azure, and amazon I haven't tried yet. Mostly I would probably test with images from the tests in the software repository, though these are limited as they these are just a few examples of course.
