### 2022-07-29 #4328 @rlskoeser
@step21 I tested against google because I already had a project setup that I was able to create credentials for. I don't have projects and accounts with the others and was holding off setting them up until I have a block of time that I'm confident I'll be able to get everything set up and tested.

I tested with an image from a project I'm currently working on, [Princeton Geniza Project](https://geniza.princeton.edu/en/) because I'm genuinely interested to know how well these services can do with this content (handwritten medieval content in Hebrew and Arabic script and a variety of languages).
