### 2022-07-29 caltechlibrary/handprint#40 @rlskoeser
JOSS review — readme should include statement of need & audience
per JOSS review checklist, the documentation in your readme should include:

> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

You can probably adapt the second paragraph from the statement of need in your JOSS paper.

Also, in both places I think you should clarify audience — you say in the paper that people can use handprint without writing any code, but it seems like people will still need to be fairly technical to get this set up and configured.
