### 2022-07-29 #4328 @rlskoeser
Code documentation doesn't include a statement of need; the statement of need in the JOSS paper should clarify the target audience.

https://github.com/caltechlibrary/handprint/issues/40
