### 2022-07-29 #4328 @rlskoeser
There is thorough documentation for installing, configuring, and using handprint from the command line. However, I'm not finding any code-level documentation.

@mhucka in the JOSS paper under the statement of need, you say that handprint could be used in "scripts as part of automated workflows." Is it your expectation that anyone using it that way would call handprint from the command line? If I'm working in python, could I use functionality from handprint as a code library without resorting to a command line call? Does handprint expose any kind of reliable API that would be safe/appropriate/reliable to use this way, or is the code (or part of it) internal and shouldn't be used that way?

My sense of your options is that you can either:
- clarify the target audience and usage to make it clear that handprint _only_ exposes a CLI and is not intended to be used as a python library
- add python api documentation and example usage for the methods that someone would need if they wanted to use handprint functionality in other python code.

Maybe @step21 or @danielskatz will have thoughts on other options here.

I'm probably thinking about it this way because as a python developer, I'd want to call it from my code for anything automated / bulk operation; but maybe I'm not your target audience! :-) If you don't want to expose a python API, it would be nice to add links to other existing tools/libraries to use instead for that kind of work. I'm trying to understand what happens after someone uses handprint to evaluate options: once they determine the best solution, can they keep using handprint or do they need to / should they switch to a different tool?
