### 2022-08-08 #4328 @step21
@rlskoeser I also think it looks more like it is meant to be automated via shellscript (or Python calling a shellscript)

I tested it with Azure for now on the IAM Handwriting Database. (https://fki.tic.heia-fr.ch/databases/iam-handwriting-database) - the functionality works as far asI can tell. (with sporadic checks of results)

It is not especially fast, but as far as I can see there are also no claims to that end, and it might depend on the service and account. This could probably be sped with parallel processing, if the service allows it.
In addition, I found some caveats mentioned in the docs which might be useful to put in a more prominent place:

```
If the input has multiple pages, only the first page/image is used; the rest (if any) are ignored.

    The Amazon Rekognition API will return [at most 50 words in an image](https://docs.aws.amazon.com/rekognition/latest/dg/limits.html).

    The Microsoft Azure API will only detect a maximum of [300 lines of text per page](https://docs.microsoft.com/en-us/azure/cognitive-services/computer-vision/concept-recognizing-text).

    Some services have different file size restrictions depending on the format of the file, but Handprint always uses the same limit for all files for a given service. This is a code simplification.
```
Especially the Amazon limitation seems kind of low to me.
