### 2022-08-08 #4328 @danielskatz
> Not sure whether or not to check off the automated tests — there are some python tests, but it's a bit hard to tell which parts of the code they cover. There is no CI set up as far as I can tell, and no documentation on how to run the tests, although I was able to run them fairly easily with pytest. (It would probably be straightforward to set up GitHub Actions to install the app and run pytest; also easy to document running them, maybe in the contributing doc.)
>
> I looked at the [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) and it doesn't seem like this quite fits in any of those categories — not automated CI, not documented; don't see any instructions on how to test expected functionality manually with sample input. Also not sure how to judge if the tests cover the "core functionality" of the application.
>
> @danielskatz what do you advise?

It looks like a bit more is needed from @mhucka. As you say, JOSS does not require CI and automated tests, but the manual tests do need to be clear to understand what they do and what is being tested.
