### 2022-08-22 caltechlibrary/handprint#38 @mhucka
Thanks for the comments.

Regarding workflows: the intention is
* People can use it in shell scripts or Python scripts (the latter using subprocesses)
* If desired, Python programmers could also extract the code for interacting with a desired service and use it as a starting point for their own programs.

Regarding people using it for automated workflows: in the past, I've had a couple of people tell me they're using it in workflows. So, yes, people seem to be using it directly.

Regarding the state of the field: I have not found any other tools that allow you to run the same input against multiple HTR services. As far as I'm aware, there are only individual sample applications, such as [Amazon's sample program Textractor](https://github.com/aws-samples/amazon-textract-textractor).

Regarding the original question by @step21, _there is no review of the "state of the field" in the paper_: that's not actually a requirement for JOSS papers, as far as I'm aware.
