### 2022-08-22 #4328 @mhucka
Regarding [@step21's comment](https://github.com/openjournals/joss-reviews/issues/4328#issuecomment-1208168096) about _"Especially the Amazon limitation seems kind of low to me"_: that's from Amazon's service – there is nothing I can do about that. I tried to make it clear that it's the API that returns those results.
