### 2022-08-22 #4328 @mhucka
I appreciate people's efforts and comments.

Regarding comparison to the field: doing a proper exploration of the state of the field today would take hours of work, which unfortunately I don't have. The last time I looked, there were no comparable tools, which is why there is no comparison in the paper currently.

If the need for a review of the state of the field is necessary, then I'm afraid I'm going to have to shelve this, because I regret I simply don't have the time.
