### 2022-08-22 #4328 @danielskatz
@mhucka - one of the JOSS review criteria is "State of the field: Do the authors describe how this software compares to other commonly-used packages?"  If you think there are no other commonly-used packages that are comparable, you might just say so in the paper.
