### 2022-09-15 #4328 @step21
> 👋 @mhucka - I think we should mark this as withdrawn, which I will do. If you decide to resubmit, please mention this issue (#4328) when you do so, and we can see if the same reviewers are available.
>
> 👋 @step21 & @rlskoeser, I'm sorry that your effort will not lead to a JOSS publication at this point, but thanks very much for your work in any case.

No worries, it was a pleasure. A good rest of the week to everyone!
